# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "North Bound Interface (NBI) REST API exposed by voltha-go-controller (VGC).\nIt provides configuration and operational endpoints for subscribers, bandwidth\nprofiles, IGMP proxy, multicast, flows, devices, ports, meters, groups, DHCP\nallocations, MAC learning and the OLT flow service.",
        "title": "VOLTHA Go Controller NBI REST API",
        "contact": {
            "name": "Open Networking Foundation",
            "url": "https://opennetworking.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "basePath": "/vgc/v1",
    "paths": {
...
        "/devices/{olt_of_id}/ports": {
            "get": {
                "description": "Retrieve the list of ports for a specific device.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get ports for a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device (OLT) OpenFlow identifier",
                        "name": "olt_of_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.DevicePortEntry"
                        }
                    }
                }
            }
        },
...
        "/flows/{deviceId}": {
            "get": {
                "description": "Retrieve all flows, flows for a specific device, or a single flow by device and flow id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flows"
                ],
                "summary": "Get flows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device identifier",
                        "name": "deviceId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.FlowEntry"
                        }
                    }
                }
            }
        },
        "/flows/{deviceId}/{flowId}": {
            "get": {
                "description": "Retrieve all flows, flows for a specific device, or a single flow by device and flow id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flows"
                ],
                "summary": "Get flows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device identifier",
                        "name": "deviceId",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Flow identifier",
                        "name": "flowId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.FlowEntry"
                        }
                    }
                }
            }
        },
...
        "/groups/{id}": {
            "get": {
                "description": "Retrieve a specific group by its identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "Get group by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.GroupList"
                        }
                    },
                    "404": {
                        "description": "Group not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "onosnbi.Bucket": {
            "type": "object",
            "properties": {
                "bucketId": {
                    "type": "integer"
                },
                "bytes": {
                    "type": "integer"
                },
                "packets": {
                    "type": "integer"
                },
                "treatment": {
                    "$ref": "#/definitions/onosnbi.Treatment"
                },
                "type": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "onosnbi.Device": {
            "type": "object",
            "properties": {
                "annotations": {
                    "$ref": "#/definitions/onosnbi.DeviceAnnotations"
                },
                "available": {
                    "type": "boolean"
                },
                "chassisId": {
                    "type": "string"
                },
                "driver": {
                    "type": "string"
                },
                "humanReadableLastUpdate": {
                    "type": "string"
                },
                "hw": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastUpdate": {
                    "type": "string"
                },
                "mfr": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "serial": {
                    "type": "string"
                },
                "sw": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "onosnbi.DeviceAnnotations": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "managementAddress": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                }
            }
        },
...
        "onosnbi.DevicePortEntry": {
            "type": "object",
            "properties": {
                "device": {
                    "$ref": "#/definitions/onosnbi.Device"
                },
                "ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onosnbi.Port"
                    }
                }
            }
        },
        "onosnbi.Flow": {
            "type": "object",
            "properties": {
                "appId": {
                    "type": "string"
                },
                "bytes": {
                    "type": "integer"
                },
                "deviceId": {
                    "type": "string"
                },
                "groupId": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "isPermanent": {
                    "type": "boolean"
                },
                "lastSeen": {
                    "type": "integer"
                },
                "life": {
                    "type": "integer"
                },
                "liveType": {
                    "type": "string"
                },
                "packets": {
                    "type": "integer"
                },
                "priority": {
                    "type": "integer"
                },
                "selector": {
                    "$ref": "#/definitions/onosnbi.SelectorInfo"
                },
                "state": {
                    "type": "string"
                },
                "tableId": {
                    "type": "integer"
                },
                "tableName": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer"
                },
                "treatment": {
                    "$ref": "#/definitions/onosnbi.TreatmentInfo"
                }
            }
        },
        "onosnbi.FlowEntry": {
            "type": "object",
            "properties": {
                "flows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onosnbi.Flow"
                    }
                }
            }
        },
        "onosnbi.GroupList": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onosnbi.GroupsInfo"
                    }
                }
            }
        },
        "onosnbi.GroupsInfo": {
            "type": "object",
            "properties": {
                "appCookie": {
                    "type": "string"
                },
                "appId": {
                    "type": "string"
                },
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onosnbi.Bucket"
                    }
                },
                "bytes": {
                    "type": "integer"
                },
                "deviceId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "life": {
                    "type": "integer"
                },
                "packets": {
                    "type": "integer"
                },
                "referenceCount": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "onosnbi.Instructions": {
            "type": "object",
            "properties": {
                "port": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "onosnbi.Port": {
            "type": "object",
            "properties": {
                "annotations": {
                    "$ref": "#/definitions/onosnbi.PortAnnotations"
                },
                "element": {
                    "type": "string"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "port": {
                    "type": "string"
                },
                "portSpeed": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "onosnbi.PortAnnotations": {
            "type": "object",
            "properties": {
                "adminState": {
                    "type": "string"
                },
                "portMac": {
                    "type": "string"
                },
                "portName": {
                    "type": "string"
                }
            }
        },
...
        "onosnbi.SelectorInfo": {
            "type": "object",
            "properties": {
                "criteria": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "onosnbi.Treatment": {
            "type": "object",
            "properties": {
                "deferred": {
                    "type": "array",
                    "items": {}
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onosnbi.Instructions"
                    }
                }
            }
        },
        "onosnbi.TreatmentInfo": {
            "type": "object",
            "properties": {
                "deferred": {
                    "type": "array",
                    "items": {}
                },
                "instructions": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
    }
}